GCP tests extraction
