+++ test/integration_test.go
@@ -2,7 +2,6 @@ package test_test
 
 import (
 	"bytes"
-	"context"
 	"encoding/json"
 	goErrors "errors"
 	"fmt"
@@ -22,24 +21,14 @@ import (
 	"testing"
 	"time"
 
-	"cloud.google.com/go/storage"
 	"github.com/aws/aws-sdk-go/aws"
 	"github.com/aws/aws-sdk-go/aws/session"
 	"github.com/aws/aws-sdk-go/service/dynamodb"
 	"github.com/aws/aws-sdk-go/service/s3"
 	"github.com/aws/aws-sdk-go/service/sts"
+	"github.com/gruntwork-io/go-commons/errors"
 	"github.com/gruntwork-io/go-commons/files"
 	"github.com/gruntwork-io/go-commons/version"
-	terraws "github.com/gruntwork-io/terratest/modules/aws"
-	"github.com/gruntwork-io/terratest/modules/git"
-	"github.com/hashicorp/go-multierror"
-	"github.com/hashicorp/hcl/v2"
-	"github.com/hashicorp/hcl/v2/hclwrite"
-	"github.com/stretchr/testify/assert"
-	"github.com/stretchr/testify/require"
-	"google.golang.org/api/iterator"
-
-	"github.com/gruntwork-io/go-commons/errors"
 	"github.com/gruntwork-io/terragrunt/aws_helper"
 	"github.com/gruntwork-io/terragrunt/cli"
 	"github.com/gruntwork-io/terragrunt/cli/commands"
@@ -54,12 +43,18 @@ import (
 	"github.com/gruntwork-io/terragrunt/remote"
 	"github.com/gruntwork-io/terragrunt/shell"
 	"github.com/gruntwork-io/terragrunt/util"
+	terraws "github.com/gruntwork-io/terratest/modules/aws"
+	"github.com/gruntwork-io/terratest/modules/git"
+	"github.com/hashicorp/go-multierror"
+	"github.com/hashicorp/hcl/v2"
+	"github.com/hashicorp/hcl/v2/hclwrite"
+	"github.com/stretchr/testify/assert"
+	"github.com/stretchr/testify/require"
 )
 
 // hard-code this to match the test fixture for now
 const (
-	terraformRemoteStateS3Region  = "us-west-2"
-	terraformRemoteStateGcpRegion = "eu"
+	terraformRemoteStateS3Region = "us-west-2"
 
 	testFixturePath                                               = "fixture/"
 	testFixtureHclvalidate                                        = "fixture-hclvalidate"
@@ -71,8 +66,6 @@ const (
 	testFixtureProviderCacheFilesystemMirror                      = "fixture-provider-cache/filesystem-mirror"
 	testFixtureDestroyOrder                                       = "fixture-destroy-order"
 	testFixtureCodegenPath                                        = "fixture-codegen"
-	testFixtureGcsPath                                            = "fixture-gcs/"
-	testFixtureGcsByoBucketPath                                   = "fixture-gcs-byo-bucket/"
 	testFixtureStack                                              = "fixture-stack/"
 	testFixtureGraphDependencies                                  = "fixture-graph-dependencies"
 	testFixtureOutputAll                                          = "fixture-output-all"
@@ -182,10 +175,7 @@ const (
 	testFixtureStrcontains                                        = "fixture-strcontains"
 	testFixtureInitCache                                          = "fixture-init-cache"
 	testFixtureNullValue                                          = "fixture-null-values"
-	testFixtureGcsImpersonatePath                                 = "fixture-gcs-impersonate/"
 	testFixtureS3Errors                                           = "fixture-s3-errors/"
-	testFixtureGcsNoBucket                                        = "fixture-gcs-no-bucket/"
-	testFixtureGcsNoPrefix                                        = "fixture-gcs-no-prefix/"
 	testFixtureDisabledPath                                       = "fixture-disabled-path/"
 	testFixtureNoSubmodules                                       = "fixture-no-submodules/"
 	testFixtureDisabledModule                                     = "fixture-disabled/"
@@ -194,7 +184,6 @@ const (
 	testFixtureTfTest                                             = "fixture-tftest/"
 	testCommandsThatNeedInput                                     = "fixture-commands-that-need-input"
 	testFixtureParallelStateInit                                  = "fixture-parallel-state-init"
-	testFixtureGcsParallelStateInit                               = "fixture-gcs-parallel-state-init"
 	testFixtureAssumeRole                                         = "fixture-assume-role/external-id"
 	testFixtureAssumeRoleDuration                                 = "fixture-assume-role/duration"
 	testFixtureAssumeRoleWebIdentityEnv                           = "fixture-assume-role-web-identity/env-var"
@@ -1074,47 +1063,6 @@ func TestTerragruntSetsAccessLoggingForTfSTateS3BuckeToADifferentBucketWithDefau
 
 	assert.Equal(t, s3BucketLogsName, targetLoggingBucket)
 	assert.Equal(t, remote.DefaultS3BucketAccessLoggingTargetPrefix, targetLoggingBucketPrefix)
-}
-
-func TestTerragruntWorksWithGCSBackend(t *testing.T) {
-	t.Parallel()
-
-	cleanupTerraformFolder(t, testFixtureGcsPath)
-
-	// We need a project to create the bucket in, so we pull one from the recommended environment variable.
-	project := os.Getenv("GOOGLE_CLOUD_PROJECT")
-	gcsBucketName := "terragrunt-test-bucket-" + strings.ToLower(uniqueId())
-
-	defer deleteGCSBucket(t, gcsBucketName)
-
-	tmpTerragruntGCSConfigPath := createTmpTerragruntGCSConfig(t, testFixtureGcsPath, project, terraformRemoteStateGcpRegion, gcsBucketName, config.DefaultTerragruntConfigPath)
-	runTerragrunt(t, fmt.Sprintf("terragrunt apply -auto-approve --terragrunt-non-interactive --terragrunt-config %s --terragrunt-working-dir %s", tmpTerragruntGCSConfigPath, testFixtureGcsPath))
-
-	var expectedGCSLabels = map[string]string{
-		"owner": "terragrunt_test",
-		"name":  "terraform_state_storage"}
-	validateGCSBucketExistsAndIsLabeled(t, terraformRemoteStateGcpRegion, gcsBucketName, expectedGCSLabels)
-}
-
-func TestTerragruntWorksWithExistingGCSBucket(t *testing.T) {
-	t.Parallel()
-
-	cleanupTerraformFolder(t, testFixtureGcsByoBucketPath)
-
-	// We need a project to create the bucket in, so we pull one from the recommended environment variable.
-	project := os.Getenv("GOOGLE_CLOUD_PROJECT")
-	gcsBucketName := "terragrunt-test-bucket-" + strings.ToLower(uniqueId())
-
-	defer deleteGCSBucket(t, gcsBucketName)
-
-	// manually create the GCS bucket outside the US (default) to test Terragrunt works correctly with an existing bucket.
-	location := terraformRemoteStateGcpRegion
-	createGCSBucket(t, project, location, gcsBucketName)
-
-	tmpTerragruntGCSConfigPath := createTmpTerragruntGCSConfig(t, testFixtureGcsByoBucketPath, project, terraformRemoteStateGcpRegion, gcsBucketName, config.DefaultTerragruntConfigPath)
-	runTerragrunt(t, fmt.Sprintf("terragrunt apply -auto-approve --terragrunt-non-interactive --terragrunt-config %s --terragrunt-working-dir %s", tmpTerragruntGCSConfigPath, testFixtureGcsByoBucketPath))
-
-	validateGCSBucketExistsAndIsLabeled(t, location, gcsBucketName, nil)
 }
 
 func TestTerragruntWorksWithSingleJsonConfig(t *testing.T) {
@@ -4652,21 +4600,6 @@ func createTmpTerragruntConfigContent(t *testing.T, contents string, configFileN
 	return tmpTerragruntConfigFile
 }
 
-func createTmpTerragruntGCSConfig(t *testing.T, templatesPath string, project string, location string, gcsBucketName string, configFileName string) string {
-	t.Helper()
-
-	tmpFolder, err := os.MkdirTemp("", "terragrunt-test")
-	if err != nil {
-		t.Fatalf("Failed to create temp folder due to error: %v", err)
-	}
-
-	tmpTerragruntConfigFile := util.JoinPath(tmpFolder, configFileName)
-	originalTerragruntConfigPath := util.JoinPath(templatesPath, configFileName)
-	copyTerragruntGCSConfigAndFillPlaceholders(t, originalTerragruntConfigPath, tmpTerragruntConfigFile, project, location, gcsBucketName)
-
-	return tmpTerragruntConfigFile
-}
-
 func copyTerragruntConfigAndFillPlaceholders(t *testing.T, configSrcPath string, configDestPath string, s3BucketName string, lockTableName string, region string) {
 	t.Helper()
 
@@ -4693,19 +4626,6 @@ func copyAndFillMapPlaceholders(t *testing.T, srcPath string, destPath string, p
 	if err := os.WriteFile(destPath, []byte(contents), 0444); err != nil {
 		t.Fatalf("Error writing temp file to %s: %v", destPath, err)
 	}
-}
-
-func copyTerragruntGCSConfigAndFillPlaceholders(t *testing.T, configSrcPath string, configDestPath string, project string, location string, gcsBucketName string) {
-	t.Helper()
-
-	email := os.Getenv("GOOGLE_IDENTITY_EMAIL")
-
-	copyAndFillMapPlaceholders(t, configSrcPath, configDestPath, map[string]string{
-		"__FILL_IN_PROJECT__":     project,
-		"__FILL_IN_LOCATION__":    location,
-		"__FILL_IN_BUCKET_NAME__": gcsBucketName,
-		"__FILL_IN_GCP_EMAIL__":   email,
-	})
 }
 
 // Returns a unique (ish) id we can attach to resources and tfstate files so they don't conflict with each other
@@ -5060,146 +4980,6 @@ func cleanupTableForTest(t *testing.T, tableName string, awsRegion string) {
 	client := createDynamoDbClientForTest(t, awsRegion)
 	err := terragruntDynamoDb.DeleteTable(tableName, client)
 	require.NoError(t, err)
-}
-
-// Check that the GCS Bucket of the given name and location exists. Terragrunt should create this bucket during the test.
-// Also check if bucket got labeled properly.
-func validateGCSBucketExistsAndIsLabeled(t *testing.T, location string, bucketName string, expectedLabels map[string]string) {
-	t.Helper()
-
-	remoteStateConfig := remote.RemoteStateConfigGCS{Bucket: bucketName}
-
-	gcsClient, err := remote.CreateGCSClient(remoteStateConfig)
-	if err != nil {
-		t.Fatalf("Error creating GCS client: %v", err)
-	}
-
-	// verify the bucket exists
-	assert.True(t, remote.DoesGCSBucketExist(gcsClient, &remoteStateConfig), "Terragrunt failed to create remote state GCS bucket %s", bucketName)
-
-	// verify the bucket location
-	ctx := context.Background()
-	bucket := gcsClient.Bucket(bucketName)
-	attrs, err := bucket.Attrs(ctx)
-	if err != nil {
-		t.Fatal(err)
-	}
-
-	assert.Equal(t, strings.ToUpper(location), attrs.Location, "Did not find GCS bucket in expected location.")
-
-	if expectedLabels != nil {
-		assertGCSLabels(t, expectedLabels, bucketName, gcsClient)
-	}
-}
-
-// gcsObjectAttrs returns the attributes of the specified object in the bucket
-func gcsObjectAttrs(t *testing.T, bucketName string, objectName string) *storage.ObjectAttrs {
-	t.Helper()
-
-	remoteStateConfig := remote.RemoteStateConfigGCS{Bucket: bucketName}
-
-	gcsClient, err := remote.CreateGCSClient(remoteStateConfig)
-	if err != nil {
-		t.Fatalf("Error creating GCS client: %v", err)
-	}
-
-	ctx := context.Background()
-	bucket := gcsClient.Bucket(bucketName)
-
-	handle := bucket.Object(objectName)
-	attrs, err := handle.Attrs(ctx)
-	if err != nil {
-		t.Fatalf("Error reading object attributes %s %v", objectName, err)
-	}
-	return attrs
-}
-
-func assertGCSLabels(t *testing.T, expectedLabels map[string]string, bucketName string, client *storage.Client) {
-	t.Helper()
-
-	ctx := context.Background()
-	bucket := client.Bucket(bucketName)
-
-	attrs, err := bucket.Attrs(ctx)
-	if err != nil {
-		t.Fatal(err)
-	}
-
-	var actualLabels = make(map[string]string)
-
-	for key, value := range attrs.Labels {
-		actualLabels[key] = value
-	}
-
-	assert.Equal(t, expectedLabels, actualLabels, "Did not find expected labels on GCS bucket.")
-}
-
-// Create the specified GCS bucket
-func createGCSBucket(t *testing.T, projectID string, location string, bucketName string) {
-	t.Helper()
-
-	var gcsConfig remote.RemoteStateConfigGCS
-	gcsClient, err := remote.CreateGCSClient(gcsConfig)
-	if err != nil {
-		t.Fatalf("Error creating GCS client: %v", err)
-	}
-
-	t.Logf("Creating test GCS bucket %s in project %s, location %s", bucketName, projectID, location)
-
-	ctx := context.Background()
-	bucket := gcsClient.Bucket(bucketName)
-
-	bucketAttrs := &storage.BucketAttrs{
-		Location:          location,
-		VersioningEnabled: true,
-	}
-
-	if err := bucket.Create(ctx, projectID, bucketAttrs); err != nil {
-		t.Fatalf("Failed to create GCS bucket %s: %v", bucketName, err)
-	}
-}
-
-// Delete the specified GCS bucket to clean up after a test
-func deleteGCSBucket(t *testing.T, bucketName string) {
-	t.Helper()
-
-	var gcsConfig remote.RemoteStateConfigGCS
-	gcsClient, err := remote.CreateGCSClient(gcsConfig)
-	if err != nil {
-		t.Fatalf("Error creating GCS client: %v", err)
-	}
-
-	t.Logf("Deleting test GCS bucket %s", bucketName)
-
-	ctx := context.Background()
-
-	// List all objects including their versions in the bucket
-	bucket := gcsClient.Bucket(bucketName)
-	q := &storage.Query{
-		Versions: true,
-	}
-	it := bucket.Objects(ctx, q)
-	for {
-		objectAttrs, err := it.Next()
-
-		if goErrors.Is(err, iterator.Done) {
-			break
-		}
-
-		if err != nil {
-			t.Fatalf("Failed to list objects and versions in GCS bucket %s: %v", bucketName, err)
-		}
-
-		// purge the object version
-		if err := bucket.Object(objectAttrs.Name).Generation(objectAttrs.Generation).Delete(ctx); err != nil {
-			t.Fatalf("Failed to delete GCS bucket object %s: %v", objectAttrs.Name, err)
-		}
-	}
-
-	// remote empty bucket
-	if err := bucket.Delete(ctx); err != nil {
-		t.Fatalf("Failed to delete GCS bucket %s: %v", bucketName, err)
-	}
 }
 
 func fileIsInFolder(t *testing.T, name string, path string) bool {
@@ -6804,41 +6584,6 @@ func TestTerragruntErrorWhenStateBucketIsInDifferentRegion(t *testing.T) {
 	assert.Contains(t, err.Error(), "BucketRegionError: incorrect region")
 }
 
-func TestTerragruntCheckMissingGCSBucket(t *testing.T) {
-	t.Parallel()
-
-	cleanupTerraformFolder(t, testFixtureGcsNoBucket)
-
-	// We need a project to create the bucket in, so we pull one from the recommended environment variable.
-	project := os.Getenv("GOOGLE_CLOUD_PROJECT")
-	gcsBucketName := "terragrunt-test-bucket-" + strings.ToLower(uniqueId())
-
-	stdout := bytes.Buffer{}
-	stderr := bytes.Buffer{}
-	tmpTerragruntGCSConfigPath := createTmpTerragruntGCSConfig(t, testFixtureGcsNoBucket, project, terraformRemoteStateGcpRegion, gcsBucketName, config.DefaultTerragruntConfigPath)
-	err := runTerragruntCommand(t, fmt.Sprintf("terragrunt apply -auto-approve --terragrunt-non-interactive --terragrunt-config %s --terragrunt-working-dir %s", tmpTerragruntGCSConfigPath, testFixtureGcsNoBucket), &stdout, &stderr)
-	require.Error(t, err)
-	assert.Contains(t, err.Error(), "Missing required GCS remote state configuration bucket")
-}
-
-func TestTerragruntNoPrefixGCSBucket(t *testing.T) {
-	t.Parallel()
-
-	cleanupTerraformFolder(t, testFixtureGcsNoPrefix)
-
-	// We need a project to create the bucket in, so we pull one from the recommended environment variable.
-	project := os.Getenv("GOOGLE_CLOUD_PROJECT")
-	gcsBucketName := "terragrunt-test-bucket-" + strings.ToLower(uniqueId())
-
-	defer deleteGCSBucket(t, gcsBucketName)
-
-	stdout := bytes.Buffer{}
-	stderr := bytes.Buffer{}
-	tmpTerragruntGCSConfigPath := createTmpTerragruntGCSConfig(t, testFixtureGcsNoPrefix, project, terraformRemoteStateGcpRegion, gcsBucketName, config.DefaultTerragruntConfigPath)
-	err := runTerragruntCommand(t, fmt.Sprintf("terragrunt apply -auto-approve --terragrunt-non-interactive --terragrunt-config %s --terragrunt-working-dir %s", tmpTerragruntGCSConfigPath, testFixtureGcsNoPrefix), &stdout, &stderr)
-	require.NoError(t, err)
-}
-
 func TestTerragruntNoWarningLocalPath(t *testing.T) {
 	t.Parallel()
 
@@ -7088,32 +6833,6 @@ func TestTerragruntParallelStateInit(t *testing.T) {
 	s3BucketName := "terragrunt-test-bucket-" + strings.ToLower(uniqueId())
 	lockTableName := "terragrunt-test-locks-" + strings.ToLower(uniqueId())
 	copyTerragruntConfigAndFillPlaceholders(t, originalTerragruntConfigPath, tmpTerragruntConfigFile, s3BucketName, lockTableName, "us-east-2")
-
-	runTerragrunt(t, "terragrunt run-all apply -auto-approve --terragrunt-non-interactive --terragrunt-working-dir "+tmpEnvPath)
-}
-
-func TestTerragruntGCSParallelStateInit(t *testing.T) {
-	t.Parallel()
-
-	tmpEnvPath, err := os.MkdirTemp("", "terragrunt-test")
-	if err != nil {
-		require.NoError(t, err)
-	}
-	for i := 0; i < 20; i++ {
-		err := util.CopyFolderContents(testFixtureGcsParallelStateInit, tmpEnvPath, ".terragrunt-test", nil)
-		require.NoError(t, err)
-		err = os.Rename(
-			path.Join(tmpEnvPath, "template"),
-			path.Join(tmpEnvPath, "app"+strconv.Itoa(i)))
-		require.NoError(t, err)
-	}
-
-	tmpTerragruntConfigFile := util.JoinPath(tmpEnvPath, "terragrunt.hcl")
-	project := os.Getenv("GOOGLE_CLOUD_PROJECT")
-	gcsBucketName := "terragrunt-test-bucket-" + strings.ToLower(uniqueId())
-	tmpTerragruntGCSConfigPath := createTmpTerragruntGCSConfig(t, testFixtureGcsParallelStateInit, project, terraformRemoteStateGcpRegion, gcsBucketName, config.DefaultTerragruntConfigPath)
-	err = util.CopyFile(tmpTerragruntGCSConfigPath, tmpTerragruntConfigFile)
-	require.NoError(t, err)
 
 	runTerragrunt(t, "terragrunt run-all apply -auto-approve --terragrunt-non-interactive --terragrunt-working-dir "+tmpEnvPath)
 }
